+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Added
 - Created ConfigurationFile class, used in load_configuration, save_configuration and check_configuration
+- Functions to update pysoem timeouts
 
 ## [7.4.1] - 2025-01-28
 ### Fixed